# swagger-spec/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/verifications/submit": {
      "post": {
        "operationId": "VerificationController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SumbitVerificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dokumen verifikasi berhasil dikirim dan menunggu peninjauan"
          },
          "400": {
            "description": "Payload atau file tidak valid"
          },
          "409": {
            "description": "Dokumen verifikasi jenis ini sedang diproses atau sudah disetujui"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit dokumen verifikasi (KTP/SIM/SKCK/STNK)",
        "tags": [
          "Verifications"
        ]
      }
    },
    "/api/verifications": {
      "get": {
        "operationId": "VerificationController_findAll",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "unverified",
                "pending",
                "approved",
                "rejected"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Daftar verifikasi berhasil diambil"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melihat seluruh daftar antrean verifikasi",
        "tags": [
          "Verifications"
        ]
      }
    },
    "/api/verifications/{id}": {
      "get": {
        "operationId": "VerificationController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail verifikasi ditemukan"
          },
          "404": {
            "description": "Verifikasi tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Melihat detail verifikasi berdasarkan ID",
        "tags": [
          "Verifications"
        ]
      }
    },
    "/api/verifications/{id}/review": {
      "patch": {
        "operationId": "VerificationController_review",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewVerificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status verifikasi berhasil ditinjau dan diperbarui"
          },
          "400": {
            "description": "Wajib mengisi alasan penolakan jika status REJECTED"
          },
          "404": {
            "description": "Verifikasi tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Approve atau Reject dokumen verifikasi",
        "tags": [
          "Verifications"
        ]
      }
    },
...
    "/api/vehicles": {
      "post": {
        "operationId": "VehiclesController_createVehicle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Kendaraan berhasil ditambahkan"
          },
          "403": {
            "description": "Akun mitra belum terverifikasi (statusVerification != approved)"
          },
          "409": {
            "description": "Nomor plat kendaraan sudah terdaftar"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Tambah kendaraan baru (Mitra Only)",
        "tags": [
          "Kendaraan"
        ]
      }
    },
    "/api/vehicles/me": {
      "get": {
        "operationId": "VehiclesController_getMyVehicles",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Daftar kendaraan ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Daftar kendaraan milik Mitra yang sedang login",
        "tags": [
          "Kendaraan"
        ]
      }
    },
    "/api/vehicles/{id}": {
      "get": {
        "operationId": "VehiclesController_getVehicleById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detail kendaraan ditemukan"
          },
          "404": {
            "description": "Data kendaraan tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Detail kendaraan berdasarkan ID",
        "tags": [
          "Kendaraan"
        ]
      },
      "patch": {
        "operationId": "VehiclesController_updateVehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateVehicleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Data kendaraan berhasil diperbarui"
          },
          "403": {
            "description": "Anda tidak memiliki akses ke kendaraan ini"
          },
          "404": {
            "description": "Data kendaraan tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update data kendaraan (Owner Only)",
        "tags": [
          "Kendaraan"
        ]
      },
      "delete": {
        "operationId": "VehiclesController_deleteVehicle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kendaraan berhasil dihapus"
          },
          "400": {
            "description": "Kendaraan masih terikat pada perjalanan aktif"
          },
          "404": {
            "description": "Data kendaraan tidak ditemukan"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Hapus kendaraan milik Mitra (Owner Only)",
        "tags": [
          "Kendaraan"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nebeng App API Specification",
    "description": "Dokumentasi API Backend Platform Nebeng Transportasi & Pengiriman Paket (Escrow Wallet & Dual QR Checkpoint)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Masukkan token JWT: Bearer <token>",
        "in": "header"
      }
    },
    "schemas": {
...
      "VerifiactionFileDto": {
        "type": "object",
        "properties": {
          "filePath": {
            "type": "string",
            "example": "/uploads/verifications/ktp-123.jpg"
          },
          "fileType": {
            "type": "string",
            "example": "image/jpeg",
            "description": "Format MIME type file hanya .jpg, .png, .pdf"
          }
        },
        "required": [
          "filePath",
          "fileType"
        ]
      },
      "SumbitVerificationDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "ktp",
              "sim",
              "skck",
              "stnk"
            ],
            "example": "ktp"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VerifiactionFileDto"
            }
          }
        },
        "required": [
          "type",
          "files"
        ]
      },
      "ReviewVerificationDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "approved",
              "rejected"
            ],
            "example": "approved"
          },
          "rejectionReason": {
            "type": "string",
            "example": "Foto Ktp Buram, silakan ulangi kembali"
          }
        },
        "required": [
          "status"
        ]
      },
...
      "CreateVehicleDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "motor",
              "mobil"
            ],
            "example": "motor"
          },
          "model": {
            "type": "string",
            "example": "Matic"
          },
          "plateNumber": {
            "type": "string",
            "example": "AD 1234 ABC"
          },
          "color": {
            "type": "string",
            "example": "Hitam"
          },
          "capacitySeats": {
            "type": "number",
            "example": 4,
            "description": "Kapasitas kursi penumpang"
          },
          "maxWeightCapacityKg": {
            "type": "number",
            "example": 50,
            "description": "Batas Max berat bagasi (kg)"
          }
        },
        "required": [
          "type",
          "model",
          "plateNumber",
          "color",
          "capacitySeats",
          "maxWeightCapacityKg"
        ]
      },
      "UpdateVehicleDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": [
              "motor",
              "mobil"
            ],
            "example": "motor"
          },
          "model": {
            "type": "string",
            "example": "Matic"
          },
          "plateNumber": {
            "type": "string",
            "example": "AD 1234 ABC"
          },
          "color": {
            "type": "string",
            "example": "Hitam"
          },
          "capacitySeats": {
            "type": "number",
            "example": 4,
            "description": "Kapasitas kursi penumpang"
          },
          "maxWeightCapacityKg": {
            "type": "number",
            "example": 50,
            "description": "Batas Max berat bagasi (kg)"
          }
        }
      },
...
    }
  }
}